requirements.txt
numpy>=1.26
pandas>=2.2
yfinance>=0.2
matplotlib>=3.9
requests>=2.32
python-dotenv>=1.0
PyYAML>=6.0

